Bugfixes: Show Messages again, Store deviceID correct, Update devices on reassign
The string manipulation for the Message was just reused from previous code and not working, because quotation marks were added at the beginning  and the end. To get the correct substrings the split character was changed and the quotations marks were deleted afterwards.

When giving a new deviceID by the server, there was an error in the copying process that removed the last character. Fixed now!

When the server assigns a new device that was created after the boot up of the M5Stick, the name on the display was not getting updated, because the list of devices is only send to the M5Stick on boot up.
To fix this the M5Stick does a device list request inside the reassign routine.